requirements.txt
transformers
torchvision
torchaudio
pillow
easyocr